Add About Us page Logged and NotLogged
